Move RenderTarget2D to a separate file
