+++ projects/PathosEngine/PathosEngine.vcxproj.filters
@@ -276,6 +276,9 @@
     <ClCompile Include="src\pathos\scene\scene_capture_component.cpp">
       <Filter>소스 파일</Filter>
     </ClCompile>
+    <ClCompile Include="src\pathos\render\render_target.cpp">
+      <Filter>소스 파일</Filter>
+    </ClCompile>
   </ItemGroup>
   <ItemGroup>
     <ClInclude Include="src\pathos\mesh\geometry.h">
@@ -600,6 +603,9 @@
       <Filter>헤더 파일</Filter>
     </ClInclude>
     <ClInclude Include="src\pathos\scene\scene_render_settings.h">
+      <Filter>헤더 파일</Filter>
+    </ClInclude>
+    <ClInclude Include="src\pathos\render\render_target.h">
       <Filter>헤더 파일</Filter>
     </ClInclude>
   </ItemGroup>
